pdf_reader_service/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
google-cloud-aiplatform
google-generativeai
